# docs/swagger.yaml
basePath: /
definitions:
  github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: Success
        type: string
      meta:
        $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.PaginationMeta'
      status:
        example: success
        type: string
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.CreateShiftRequest:
    properties:
      bujp_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      description:
        example: Shift pagi dari pukul 07:00 hingga 15:00
        type: string
      duration_hours:
        example: 8
        type: number
      end_time:
        example: "15:00"
        type: string
      name:
        example: Shift Pagi
        maxLength: 50
        type: string
      start_time:
        example: "07:00"
        type: string
      status:
        enum:
        - active
        - inactive
        example: active
        type: string
    required:
    - end_time
    - name
    - start_time
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.PaginationMeta:
    properties:
      has_next:
        example: true
        type: boolean
      has_previous:
        example: false
        type: boolean
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.UpdateShiftRequest:
    properties:
      bujp_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      description:
        example: Shift pagi dari pukul 07:00 hingga 15:00
        type: string
      duration_hours:
        example: 8
        type: number
      end_time:
        example: "15:00"
        type: string
      name:
        example: Shift Pagi
        maxLength: 50
        type: string
      start_time:
        example: "07:00"
        type: string
      status:
        enum:
        - active
        - inactive
        example: active
        type: string
    type: object
...
host: localhost:4000
info:
  contact:
    email: support@hrmis.com
    name: API Support
  description: HRMIS Backend API - Human Resource Management Information System
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: HRMIS API
  version: "1.0"
paths:
...
  /api/v1/shifts:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all shifts
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by BUJP ID
        in: query
        name: bujp_id
        type: string
      - description: Filter by status (active/inactive)
        in: query
        name: status
        type: string
      - description: Search by name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all shifts
      tags:
      - Shifts
    post:
      consumes:
      - application/json
      description: Create a new shift with the provided data
      parameters:
      - description: Shift data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.CreateShiftRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new shift
      tags:
      - Shifts
  /api/v1/shifts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a shift by ID
      parameters:
      - description: Shift UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a shift
      tags:
      - Shifts
    get:
      consumes:
      - application/json
      description: Retrieve a single shift by its UUID
      parameters:
      - description: Shift UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Get shift by ID
      tags:
      - Shifts
    put:
      consumes:
      - application/json
      description: Update an existing shift by ID
      parameters:
      - description: Shift UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated shift data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.UpdateShiftRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Update a shift
      tags:
      - Shifts
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"